Extract Filters.py file
