+++ minis/downloader/__init__.py
@@ -1,2 +1,3 @@
 from .downloadAllMinis import *
 from .Pages import *
+from .Filters import *
